InGame 컨트롤방식 조절가능하도록 어느정도 설정완료 아이템 기본틀 완성 SuperPower 초안완성

+++ Assets/03.Scripts/InGame/ScrollController.cs
@@ -4,7 +4,7 @@ using UnityEngine.UI;
 
 public class RefData
 {
-    public int _Value = 0;
+    public float _Value = 0;
 }
 
 [Serializable]
